fix api

+++ src/main/java/com/example/mapper/LoanSlipMapper.java
@@ -11,8 +11,8 @@ public class LoanSlipMapper implements RowMapper<LoanSlipModel>{
         try {
             LoanSlipModel loanSlipModel = new LoanSlipModel();
             loanSlipModel.setId((int) rs.getLong("id"));
-            loanSlipModel.setIdAccount(rs.getInt("idAccount"));
-            loanSlipModel.setIdBook(rs.getInt("idBook"));
+            loanSlipModel.setIdAccount((int) rs.getLong("idAccount"));
+            loanSlipModel.setIdBook((int) rs.getLong("idBook"));
             loanSlipModel.setCode(rs.getString("code"));
             loanSlipModel.setCreated_at(rs.getTimestamp("created_at"));
             loanSlipModel.setUpdated_at(rs.getTimestamp("updated_at"));
